pratt.c: store min_bp in parser struct
instead of hardcoded 0, this makes the expression parser more flexible.
for example, in C expression grammar, it described the rule for
constant-expression like this:

    constant-expression:
        conditional-expression

conditional-expression is `?:` or ternary operator, and in C operator
precedence table, it's above assignment and comma expression.

 ->    ?:                                  right to left
       = *= /= %= += -= <<= >>= &= ^= |=   right to left
       ,                                   left to right

and suppose we have this binding power table:

       ?:                                  20
       = *= /= %= += -= <<= >>= &= ^= |=   10
       ,                                    0

now if we use the previous hardcode, how can we skip the assignment
and comma expression? obviously we can't, because each time we need
to parse some expression that needs constant-expression like `#if`-
directive, it will also parse assignment and comma expression!